Feat: Add getScore function to calculate score of server with his latency, bandwidth and distance.

+++ src/main/java/com/ortega/scdn/models/Server.java
@@ -17,7 +17,7 @@ public class Server {
     private UUID id = UUID.randomUUID();
     private String name;
     private Location location;
-    private int latency;
-    private int bandwidth;
+    private double latency;
+    private double bandwidth;
 
 }
